+++ event-sourcing/event-repository/event-repository-jdbc/src/test/java/uk/gov/justice/services/eventsourcing/repository/jdbc/event/MultipleDataSourceLinkedEventRepositoryFactoryTest.java
@@ -33,10 +33,10 @@ public class MultipleDataSourceLinkedEventRepositoryFactoryTest {
 
         final DataSource dataSource = mock(DataSource.class);
 
-        final MultipleDataSourcePublishedEventRepository multipleDataSourcePublishedEventRepository = multipleDataSourcePublishedEventRepositoryFactory.create(dataSource);
+        final MultipleDataSourceEventRepository multipleDataSourceEventRepository = multipleDataSourcePublishedEventRepositoryFactory.create(dataSource);
 
-        assertThat(getValueOfField(multipleDataSourcePublishedEventRepository, "jdbcResultSetStreamer", JdbcResultSetStreamer.class), is(jdbcResultSetStreamer));
-        assertThat(getValueOfField(multipleDataSourcePublishedEventRepository, "preparedStatementWrapperFactory", PreparedStatementWrapperFactory.class), is(preparedStatementWrapperFactory));
-        assertThat(getValueOfField(multipleDataSourcePublishedEventRepository, "dataSource", DataSource.class), is(dataSource));
+        assertThat(getValueOfField(multipleDataSourceEventRepository, "jdbcResultSetStreamer", JdbcResultSetStreamer.class), is(jdbcResultSetStreamer));
+        assertThat(getValueOfField(multipleDataSourceEventRepository, "preparedStatementWrapperFactory", PreparedStatementWrapperFactory.class), is(preparedStatementWrapperFactory));
+        assertThat(getValueOfField(multipleDataSourceEventRepository, "dataSource", DataSource.class), is(dataSource));
     }
 }