+++ test-utils-event-store/test-utils-event/src/test/java/uk/gov/justice/services/test/utils/events/LinkedEventBuilderTest.java
@@ -3,7 +3,7 @@ package uk.gov.justice.services.test.utils.events;
 import static java.util.UUID.fromString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static uk.gov.justice.services.test.utils.events.PublishedEventBuilder.publishedEventBuilder;
+import static uk.gov.justice.services.test.utils.events.LinkedEventBuilder.publishedEventBuilder;
 
 import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.LinkedEvent;